ConnectionSpeed: first implemented [1/2]
This is a re-worked of Traffic monitor.
We noticed that many people actually don't want
complicated settings, so that we cleaned those codes
a little bit, but still give enough information for
user. Moreover, we fixed the size of the number so that
when the speed varies, we don't be distracted by those
numbers.

Change-Id: I8fcaf27ca19984a5a4d9b8dceb487d69e201ec67